[MODORDERS-890] - refactoring
